+++ example/bin/color_example.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:ansi_modifier/src/ansi.dart';
 
 void main(List<String> args) {
@@ -30,12 +28,4 @@ void main(List<String> args) {
 
   // Strip all Ansi modifiers.
   print('$yellowGreen, $magenta, ${magenta.clearStyle()}\n');
-
-  print('Moving the cursor:');
-  stdout.write('Hello world!');
-  stdout.write(Ansi.cursorToColumn(1));
-  stdout.write('Say hi to the ' + 'moon'.style(Ansi.yellow) + '!');
-  stdout.write(Ansi.cursorForward(10));
-  stdout.write('Hi there.');
-  print('');
 }